chore: move integration test files
