+++ notionrs/tests/mutable/data_source/mod.rs
@@ -1,3 +1,1 @@
 mod crud_data_source;
-mod list_data_source_templates;
-mod retrieve_data_source;
