+++ notionrs/tests/readonly/data_source/retrieve_data_source.rs
@@ -1,13 +1,13 @@
 mod integration_tests {
 
+    static DATA_SOURCE_ID: &str = "33da03d7-9b26-81cb-90c7-000b8fb827a8";
+
     #[tokio::test]
     async fn retrieve_data_source() -> Result<(), notionrs::Error> {
-        dotenvy::dotenv().ok();
+        dotenvy::from_path(std::path::Path::new(".env.readonly"))
+            .expect("Failed to load .env.readonly file");
 
-        let data_source_id =
-            std::env::var("NOTION_IT_DATA_SOURCE_ID").unwrap_or_else(|_| String::new());
-
-        let notion_api_key = std::env::var("NOTION_TOKEN").unwrap();
+        let notion_api_key = std::env::var("NOTION_API_KEY").unwrap();
         let client = notionrs::Client::new(notion_api_key);
 
         // # --------------------------------------------------------------------------------
@@ -16,7 +16,7 @@ mod integration_tests {
         //
         // # --------------------------------------------------------------------------------
 
-        let request = client.retrieve_data_source().data_source_id(data_source_id);
+        let request = client.retrieve_data_source().data_source_id(DATA_SOURCE_ID);
 
         let response = request.send().await?;
